Переместить описание типов
