--- src/models/basket.ts
@@ -1,5 +1,0 @@
-import { Basket } from '../types';
-
-export function emptyBasket(): Basket {
-	return { items: [] };
-}
