--- src/models/product.ts
@@ -1,10 +1,0 @@
-export type ProductId = string;
-
-export type Product = {
-	id: ProductId;
-	description: string;
-	image: string;
-	title: string;
-	category: string;
-	price: number | null;
-};
